SOY RETRASADO

+++ Modules/logmod.py
@@ -5,7 +5,8 @@ configdir   = "Config.ini"
 config.read(configdir)
 
 log = []
-logidr = None
+logdir = None
+ret = 1
 def writeonlog(thingtowrite, *otherthingstowrite):
     global log
     global logdir
@@ -18,11 +19,16 @@ def writeonlog(thingtowrite, *otherthingstowrite):
 
 def savelog():
     global log
-    with open(config.get("DIRS", "Log"), "a") as logfile:
-        logfile.writelines(log)
-        log = []
+    global logdir
+    global ret
+    if ret:
+        with open(logdir + "Log.log", "a") as logfile:
+            logfile.writelines(log)
+            log = []
 
 def createlogfile():
+    global logdir
+    global ret
     if config.get("DIRS", "logdir") == "Default":
         if not os.path.exists("logfiles/"):
             try:
@@ -30,29 +36,37 @@ def createlogfile():
                 logdir = "logfiles/"
             except:
                 logdir = "~/.invproy/logfiles/"
-                if not os.path.exists(logdir)
+                if not os.path.exists(logdir):
                     try:
                         os.makedirs(logdir)
+                    except:
+                        print("No se ha podido crear {}".format(logdir))
+                        ret = 0
     else:
         logdir = config.get("DIRS", "logdir")
-
-    nlogfiles = int(len(os.listdir(logdir)))
-    if nlogfiles >= int(config.get("DIRS", "Maxlogs")):
-        while nlogfiles > int(config.get("DIRS", "Maxlogs")):
-            #Aqui pones que borre el archivo mas viejo
-            nlogfiles -= 1
-            log.append("Borrado: " + str(min(os.listdir("logfiles/")))+ "\n")
+        if not os.path.exists(logdir):
             try:
-                os.remove("logfiles/" + min(os.listdir("logfiles/")))
-            except OSError:
-                print("\033[31mError de I/O en {}, borrar la carpeta de logfiles\033[00m".format(str(OSError.filename)))
+                os.makedirs(logdir)
             except:
-                raise
-    try:
-        newlogfilename = logdir + time.strftime("%y%m%d%H%M%S") + " " +  config.get("DIRS", "Log")
+                ret = 0
+    if ret:
+        nlogfiles = int(len(os.listdir(logdir)))
+        if nlogfiles >= int(config.get("DIRS", "Maxlogs")):
+            while nlogfiles > int(config.get("DIRS", "Maxlogs")):
+                #Aqui pones que borre el archivo mas viejo
+                nlogfiles -= 1
+                log.append("Borrado: " + str(min(os.listdir("logfiles/")))+ "\n")
+                try:
+                    os.remove("logfiles/" + min(os.listdir("logfiles/")))
+                except OSError:
+                    print("\033[31mError de I/O en {}, borrar la carpeta de logfiles\033[00m".format(str(OSError.filename)))
+                except:
+                    raise
         try:
-            os.rename("Log.log", newlogfilename)
+            newlogfilename = logdir + time.strftime("%y%m%d%H%M%S") + ".log"
+            try:
+                os.rename("Log.log", newlogfilename)
+            except:
+                print('Ojo cuidao que no se ha podio renombrar <Log.log>')
         except:
-            print('Ojo cuidao que no se ha podio renombrar <Log.log>')
-    except:
-        pass
+            pass
